# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "名单管理系统",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/name-list",
    "paths": {
        "/item": {
            "get": {
                "description": "名单项查找功能, 通过code, is_valid, 分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "名单项"
                ],
                "summary": "名单项查找功能",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Code",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否生效",
                        "name": "isValid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "PageIndex",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "PageSize",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SortedField",
                        "name": "sortedField",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正常回包, 回复查询成功的名单类型数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpx.JSONResultPaged"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.ListItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "名单项创建功能",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "名单项"
                ],
                "summary": "名单项创建功能",
                "parameters": [
                    {
                        "description": "名单属性",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/list_item.CreateListItemInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正常回包, 回复创建成功的名单类型数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpx.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.ListItem"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/item/{id}": {
            "delete": {
                "description": "名单项删除功能, 通过list_item_id删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "名单项"
                ],
                "summary": "名单项删除功能",
                "parameters": [
                    {
                        "minLength": 1,
                        "type": "string",
                        "description": "名单项id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpx.JSONResult"
                        }
                    }
                }
            },
            "patch": {
                "description": "名单项修改功能, 通过code, is_valid, 分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "名单项"
                ],
                "summary": "名单项修改功能",
                "parameters": [
                    {
                        "minLength": 1,
                        "type": "string",
                        "description": "名单项id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "名单属性",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/list_item.UpdateListItemInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正常回包, 回复更新成功的名单类型数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpx.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.ListItem"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/type": {
            "get": {
                "description": "名单类型查找功能, 通过code, is_valid, 分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "名单类型"
                ],
                "summary": "名单类型查找功能",
                "parameters": [
                    {
                        "type": "string",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "isValid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sortedField",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正常回包, 回复查询成功的名单类型数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpx.JSONResultPaged"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.ListType"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "名单类型创建功能",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "名单类型"
                ],
                "summary": "名单类型创建功能",
                "parameters": [
                    {
                        "description": "名单属性",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/list_type.CreateListTypeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正常回包, 回复创建成功的名单类型数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpx.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.ListType"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/type/{id}": {
            "delete": {
                "description": "名单类型删除功能, 通过list_type_id删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "名单类型"
                ],
                "summary": "名单类型删除功能",
                "parameters": [
                    {
                        "minLength": 1,
                        "type": "string",
                        "description": "名单类型id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正常回包, 回复删除成功的名单类型数据",
                        "schema": {
                            "$ref": "#/definitions/httpx.JSONResult"
                        }
                    }
                }
            },
            "patch": {
                "description": "名单类型修改功能",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "名单类型"
                ],
                "summary": "名单类型修改功能",
                "parameters": [
                    {
                        "minLength": 1,
                        "type": "string",
                        "description": "名单类型id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "名单属性",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/list_type.UpdateListTypeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正常回包, 回复更新成功的名单类型数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpx.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.ListType"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "httpx.JSONResult": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "回包code，表明是否正确，在code == 0时，表明服务正常",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "type": "object"
                },
                "message": {
                    "description": "回报message，在code != 0时，展示给前端",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "httpx.JSONResultPaged": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "回包code，表明是否正确，在code == 0时，表明服务正常",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "type": "object"
                },
                "message": {
                    "description": "回报message，在code != 0时，展示给前端",
                    "type": "string",
                    "example": "success"
                },
                "total": {
                    "description": "总数量",
                    "type": "integer"
                }
            }
        },
        "list_item.CreateListItemInfo": {
            "type": "object",
            "required": [
                "code",
                "is_valid",
                "namespace",
                "values"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "extra": {
                    "type": "object",
                    "additionalProperties": true
                },
                "is_valid": {
                    "type": "boolean"
                },
                "namespace": {
                    "type": "string"
                },
                "values": {
                    "description": "与list type中的fields一一对应的",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "list_item.UpdateListItemInfo": {
            "type": "object",
            "properties": {
                "extra": {
                    "description": "可自定义的结构, 不管控",
                    "type": "object",
                    "additionalProperties": true
                },
                "is_valid": {
                    "description": "是否生效",
                    "type": "boolean"
                }
            }
        },
        "list_type.CreateListTypeReq": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "名单类型编码",
                    "type": "string",
                    "example": "telephone"
                },
                "description": {
                    "description": "描述",
                    "type": "string",
                    "example": "description"
                },
                "fields": {
                    "description": "这类名单的值被构建的字段",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "telephone",
                        "id_card"
                    ]
                },
                "is_valid": {
                    "description": "是否生效",
                    "type": "boolean"
                },
                "namespace": {
                    "description": "命名空间",
                    "type": "string",
                    "example": "anti-fraud"
                }
            }
        },
        "list_type.UpdateListTypeReq": {
            "type": "object",
            "required": [
                "description",
                "is_valid"
            ],
            "properties": {
                "description": {
                    "description": "描述",
                    "type": "string",
                    "example": "description"
                },
                "is_valid": {
                    "description": "是否生效",
                    "type": "boolean"
                }
            }
        },
        "models.ListItem": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "名单类型编码",
                    "type": "string",
                    "example": "telephone__red"
                },
                "extra": {
                    "description": "可自定义的结构, 不管控",
                    "type": "object",
                    "additionalProperties": true
                },
                "id": {
                    "description": "主键id",
                    "type": "string",
                    "example": "60d2b17f70d9d2f0db53f866"
                },
                "is_valid": {
                    "description": "是否生效",
                    "type": "boolean"
                },
                "multi_value": {
                    "description": "多项值列表，如果名单是由多个字段构成的，可一一罗列出每个字段的值，如：[{\"key\":\"field1\",\"value\":\"value1\"},{\"key\":\"field2\",\"value\":\"value2\"}]",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MultiValueItem"
                    }
                },
                "namespace": {
                    "description": "命名空间",
                    "type": "string",
                    "example": "anti-fraud"
                },
                "value": {
                    "description": "名单项的值",
                    "type": "string",
                    "example": "13333333333"
                }
            }
        },
        "models.ListType": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "名单类型编码",
                    "type": "string",
                    "example": "telephone__red"
                },
                "description": {
                    "description": "描述",
                    "type": "string",
                    "example": "descriptions"
                },
                "fields": {
                    "description": "这类名单的值被构建的字段",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "telephone",
                        "id_card"
                    ]
                },
                "id": {
                    "description": "主键id",
                    "type": "string",
                    "example": "60d2b17f70d9d2f0db53f866"
                },
                "is_valid": {
                    "description": "是否生效",
                    "type": "boolean"
                },
                "namespace": {
                    "description": "命名空间",
                    "type": "string",
                    "example": "anti-fraud"
                }
            }
        },
        "models.MultiValueItem": {
            "type": "object",
            "properties": {
                "key": {
                    "description": "field name",
                    "type": "string",
                    "example": "telephone"
                },
                "value": {
                    "description": "value",
                    "type": "string",
                    "example": "13333333333"
                }
            }
        },
...
    }
}